requirements.txt
telethon
python-dotenv
yfinance
apscheduler
pytz
feedparser
requests
googletrans